+++ race_control.go
@@ -436,9 +436,7 @@ func (rc *RaceControl) OnSessionUpdate(sessionInfo udp.SessionInfo) (bool, error
 
 // OnEndSession is called at the end of every session.
 func (rc *RaceControl) OnEndSession(sessionFile udp.EndSession) error {
-	logrus.Debug("On end session")
-
-	if rc.currentTimeAttackEvent != nil {
+	if rc.currentTimeAttackEvent != nil && Premium() {
 		filename := filepath.Base(string(sessionFile))
 
 		err := rc.addFileToTimeAttackEvent(filename)